backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
email-validator>=2.0.0
motor==3.7.0
pymongo==4.13.0
pydantic==2.10.6
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
aiofiles==24.1.0
